Address #7
There are no short-term plans to make this library more convenient. If this functionality is going to get easier, I would do it in the language. On the prioritization list, this is not super high. We are not adding expressiveness here, only changing how we say the same things. In other words, other things need attention first.